Add Server-Side Timeout for Locks
In some cases locks can be orphaned on the server due to things like application crashes or network interruptions. When the underlying TCP connections are left open a lock can remain on the server despite the calling application not longer being active. This change introduces a timeout for locks via the `WithLockTimeoutSeconds` option, which sets the `wait_timeout` session variable, allowing MySQL to disconnect and release the lock after a period of inactivity.